+++ sdk/identity/azure-identity/src/workload_identity_credential.cpp
@@ -30,20 +30,17 @@ constexpr auto AzureFederatedTokenFileEnvVarName = "AZURE_FEDERATED_TOKEN_FILE";
 } // namespace
 
 WorkloadIdentityCredential::WorkloadIdentityCredential(
-    std::string const& tenantIdOptions,
-    std::string const& clientIdOptions,
-    std::string const& authorityHostOptions,
-    std::string const& tokenFilePathOptions,
-    std::vector<std::string> additionallyAllowedTenants,
-    Core::Credentials::TokenCredentialOptions const& options)
-    : TokenCredential("WorkloadIdentityCredential"),
-      m_clientCredentialCore(tenantIdOptions, authorityHostOptions, additionallyAllowedTenants)
+    WorkloadIdentityCredentialOptions const& options)
+    : TokenCredential("WorkloadIdentityCredential"), m_clientCredentialCore(
+                                                         options.TenantId,
+                                                         options.AuthorityHost,
+                                                         options.AdditionallyAllowedTenants)
 {
   std::string tenantId;
   std::string clientId;
   std::string authorityHost;
 
-  if (tenantIdOptions.empty())
+  if (options.TenantId.empty())
   {
     tenantId = Environment::GetVariable(AzureTenantIdEnvVarName);
     if (tenantId.empty())
@@ -52,7 +49,7 @@ WorkloadIdentityCredential::WorkloadIdentityCredential(
           "No tenant ID specified. Check pod configuration or set TenantID in the options.");
     }
   }
-  if (clientIdOptions.empty())
+  if (options.ClientId.empty())
   {
     clientId = Environment::GetVariable(AzureClientIdEnvVarName);
     if (clientId.empty())
@@ -61,7 +58,7 @@ WorkloadIdentityCredential::WorkloadIdentityCredential(
           "No client ID specified. Check pod configuration or set ClientID in the options.");
     }
   }
-  if (authorityHostOptions.empty())
+  if (options.AuthorityHost.empty())
   {
     authorityHost = Environment::GetVariable(AzureAuthorityHostEnvVarName);
     if (authorityHost.empty())
@@ -69,7 +66,7 @@ WorkloadIdentityCredential::WorkloadIdentityCredential(
       authorityHost = _detail::ClientCredentialCore::AadGlobalAuthority;
     }
   }
-  if (tokenFilePathOptions.empty())
+  if (options.TokenFilePath.empty())
   {
     m_tokenFilePath = Environment::GetVariable(AzureFederatedTokenFileEnvVarName);
     if (m_tokenFilePath.empty())
@@ -80,7 +77,7 @@ WorkloadIdentityCredential::WorkloadIdentityCredential(
   }
 
   m_clientCredentialCore = Azure::Identity::_detail::ClientCredentialCore(
-      tenantId, authorityHost, additionallyAllowedTenants);
+      tenantId, authorityHost, options.AdditionallyAllowedTenants);
   m_tokenCredentialImpl = std::make_unique<TokenCredentialImpl>(options);
   m_requestBody
       = std::string(
@@ -92,27 +89,43 @@ WorkloadIdentityCredential::WorkloadIdentityCredential(
 }
 
 WorkloadIdentityCredential::WorkloadIdentityCredential(
-    WorkloadIdentityCredentialOptions const& options)
-    : WorkloadIdentityCredential(
-        options.TenantId,
-        options.ClientId,
-        options.AuthorityHost,
-        options.TokenFilePath,
-        options.AdditionallyAllowedTenants,
-        options)
-{
-}
-
-WorkloadIdentityCredential::WorkloadIdentityCredential(
     Core::Credentials::TokenCredentialOptions const& options)
-    : WorkloadIdentityCredential(
-        WorkloadIdentityCredentialOptions{}.TenantId,
-        WorkloadIdentityCredentialOptions{}.ClientId,
-        WorkloadIdentityCredentialOptions{}.AuthorityHost,
-        WorkloadIdentityCredentialOptions{}.TokenFilePath,
-        WorkloadIdentityCredentialOptions{}.AdditionallyAllowedTenants,
-        options)
 {
+  std::string tenantId = Environment::GetVariable(AzureTenantIdEnvVarName);
+  if (tenantId.empty())
+  {
+    throw std::runtime_error(
+        "No tenant ID specified. Check pod configuration or set TenantID in the options.");
+  }
+  std::string clientId = Environment::GetVariable(AzureClientIdEnvVarName);
+  if (clientId.empty())
+  {
+    throw std::runtime_error(
+        "No client ID specified. Check pod configuration or set ClientID in the options.");
+  };
+  std::string authorityHost = Environment::GetVariable(AzureAuthorityHostEnvVarName);
+  if (authorityHost.empty())
+  {
+    authorityHost = _detail::ClientCredentialCore::AadGlobalAuthority;
+  }
+
+  m_tokenFilePath = Environment::GetVariable(AzureFederatedTokenFileEnvVarName);
+  if (m_tokenFilePath.empty())
+  {
+    throw std::runtime_error(
+        "No token file specified. Check pod configuration or set TokenFilePath in the options.");
+  }
+
+  m_clientCredentialCore = Azure::Identity::_detail::ClientCredentialCore(
+      tenantId, authorityHost, std::vector<std::string>());
+  m_tokenCredentialImpl = std::make_unique<TokenCredentialImpl>(options);
+  m_requestBody
+      = std::string(
+            "grant_type=client_credentials"
+            "&client_assertion_type="
+            "urn%3Aietf%3Aparams%3Aoauth%3Aclient-assertion-type%3Ajwt-bearer" // cspell:disable-line
+            "&client_id=")
+      + Url::Encode(clientId);
 }
 
 WorkloadIdentityCredential::~WorkloadIdentityCredential() = default;