+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 
- - Add support for reading the tenant id, client id, and the token file path for `WorkloadIdentityCredential` from the environment variables.
+- Add support for reading the tenant id, client id, and the token file path for `WorkloadIdentityCredential` from the environment variables.
 
 ### Breaking Changes